Remove user ID extraction from token in ThuongHieuController
The code changes involve removing the extraction of the user ID from the token in the `ThuongHieuController.cs` file within the `API.Controllers.SanPham_Controller` namespace. Specifically, the lines that extract the token from the `Authorization` header and retrieve the user ID using `_jwtService.GetMaTaiKhoanFromToken(token)` have been removed. This also includes the conditional check that returns an unauthorized response if the user ID is null.

The code previously included logic to extract a user ID from a JWT token found in the `Authorization` header of the HTTP request. This user ID was then used to verify the user's identity. If the user ID could not be retrieved, an unauthorized response was returned. These lines have now been removed, indicating that the user ID extraction and verification process is no longer required in this part of the code.